refactor: Buyer 객체 생성 및 책임 분리
