+++ src/main/java/domain/LottoFactory.java
@@ -13,12 +13,6 @@ import java.util.stream.IntStream;
 
 public class LottoFactory {
 
-    private final List<Lotto> lottos;
-
-    public LottoFactory(Money money) {
-        this.lottos = createLottos(money);
-    }
-
     public List<Lotto> createLottos(Money money) {
         List<Lotto> lottos = new ArrayList<>();
 
@@ -35,28 +29,9 @@ public class LottoFactory {
             .collect(Collectors.toList());
         Collections.shuffle(numbers);
 
-        return numbers.subList(0, LOTTO_SIZE);
-    }
+        List<Integer> selectedNumbers = numbers.subList(0, LOTTO_SIZE);
+        Collections.sort(selectedNumbers);
 
-    public String createResult() {
-        String result = "";
-        for (Lotto lotto : lottos) {
-            result += lotto.toString() + "\n";
-        }
-        return result;
-    }
-
-    public Map<LottoRank, Integer> countLottos(WinningLotto winningLotto) {
-        Map<LottoRank, Integer> result = LottoRank.createLottoRankCounter();
-
-        for (Lotto lotto : lottos) {
-            LottoRank matchedLotto = lotto.compareLotto(winningLotto);
-            result.put(matchedLotto, result.get(matchedLotto) + 1);
-        }
-        return result;
-    }
-
-    public int getLottoSize() {
-        return lottos.size();
+        return selectedNumbers;
     }
 }
